+++ test/fixtures/generatekeyaddrvector.py
@@ -1,5 +1,5 @@
-from ethereum import tester
-from ethereum import utils
+from vapory import tester
+from vapory import utils
 from bitcoin import ecdsa_sign, ecdsa_raw_sign, ecdsa_raw_recover, decode_sig
 import json